pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gtnh-agent"
version = "0.1.0"
description = "LLM-based GTNH agent."
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Knoah", email = "noah.meserve@icloud.com" }]

# Runtime dependencies go here
dependencies = [
  "pyyaml>=6.0",            # YAML config loading (env, semantics, curricula, virtues)
  "networkx",
  "pydantic",
  "rich>=13.0",             # TUI dashboard + pretty logs
  "typing-extensions>=4.0", # Future-proof typed code on Python 3.10
  # Add more as you solidify choices (e.g. fastapi, numpy, etc.)
]

[project.optional-dependencies]
dev = [                      # `pip install -e .[dev]`
  "pytest>=7.0",            # Test runner
  "pytest-cov>=4.0",        # Coverage reports
]
lint = [                     # `pip install -e .[lint]`
  "ruff>=0.6",              # Fast linter
  "mypy>=1.8",              # Type checker
]

[project.urls]
homepage = "https://example.com/gtnh-agent"
repository = "https://example.com/gtnh-agent.git"

[tool.setuptools]
package-dir = { "" = "src" } # Root package dir is src/ (src layout)

[tool.setuptools.packages.find]
where = ["src"]              # Look for packages only under src/
include = ["*"]              # Include all packages found there
namespaces = false           # Not using namespace packages

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
  "tests",
]
python_files = [
  "test_*.py",
]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src"]

[tool.mypy]
python_version = "3.10"
mypy_path = ["src"]
ignore_missing_imports = true
strict_optional = true

